pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lcz"
version = "0.1.1"
authors = [
    {name="Marcel Claramunt", email="marcel@moveread.com"}
]
description = "Simplified Leela Chess Zero (aka lc0 aka lcz) with Python"
dependencies = [
    "meson==1.2.2",
    "lczero-bindings",
    "ramda"
]
requires-python = ">=3.10"
readme = {file="README.md", content-type="text/markdown"}

[project.urls]
repo = "https://github.com/moveread/lc0"
